Use a Concern rather than a file override to handle the Process routing
